requirements.txt
Flask>=3.0,<4.0
Pillow>=10.0,<11.0
